fix RpcHandlers with bounded channel

+++ client/service/src/lib.rs
@@ -106,7 +106,7 @@ impl RpcHandlers {
 		json_query: &str,
 	) -> Result<(String, tokio::sync::mpsc::Receiver<String>), JsonRpseeError> {
 		self.0
-			.raw_json_request(json_query, usize::MAX)
+			.raw_json_request(json_query, u32::MAX as usize)
 			.await
 			.map(|(method_res, recv)| (method_res.result, recv))
 	}
